Refactored storage engine to improve deployment injection
