+++ src/index.ts
@@ -3,7 +3,7 @@ import { IEncryptionLayer, EncryptionLayerPGP, EncryptionLayerAES } from './encr
 import { IdentityManager, ShareableIdentity } from './indentityManager';
 import { IKeys, IKeysGenerator, KeysGeneratorPGP } from './keysGenerator';
 import FactoryIdentity from './factoryIdentity/FactoryIdentity';
-import { StorageEngine, IStorageEngine } from './storageEngine';
+import { StorageEngine, Storage, IStorageEngine } from './storageEngine';
 import {FactoryInteraction} from './contractIntegration';
 import Web3Provider from './contractIntegration/interaction/Web3Provider';
 import { DocumentSharing, IDocumentSharing } from './documentSharing';
@@ -20,6 +20,7 @@ export {
     IKeys,
     IKeysGenerator,
     StorageEngine,
+    Storage,
     IStorageEngine,
     Web3Provider,
     DocumentSharing,
