+++ src/storageEngine/StorageEngine.ts
@@ -1,160 +1,38 @@
+import 'reflect-metadata';
+import { container } from 'tsyringe';
+import { constructor } from 'tsyringe/dist/typings/types';
+import IPFSEngine from './IPFSEngine';
 import IStorageEngine from './IStorageEngine';
-import { injectable } from 'tsyringe';
-import axios, { AxiosInstance } from 'axios';
-import { ecsign } from 'ethereumjs-util';
-import FormData from 'form-data';
-import { Buffer } from 'buffer';
-import { IIpfsConstructor } from './IPFS/IIPFSConstructor';
-import { IIpfsDelete } from './IPFS/IIPFSDelete';
-import { IIpfsSignature } from './IPFS/IIPFSSignature';
-import { IIpfsDocument } from './IPFS/IIPFSDocuments';
-import { IIpfsUpdate } from './IPFS/IIPFSUpdate';
-import { IIpfsSave } from './IPFS/IIPFSSave';
+import Storage from './Storage';
 
 /**
- * Storage engine to save, update, get and delete files from IPFS.
- * You can inject your preferred storage engine.
+ * Storage engine to save, update, get and delete files based on the injected engine
  */
-@injectable()
-class StorageEngine implements IStorageEngine {
-    private instance: AxiosInstance;
+class StorageEngine {
+    private storageEngine: IStorageEngine;
 
     /**
-     * Storage engine constructor
-     * Using to initialize the axios instance
-     * @param {IIpfsConstructor} options - Connection options
+     * Create new container and inject storage engine
+     * @param {constructor<unknown>} engine - Storage engine type
      */
-    constructor (options: IIpfsConstructor) {
-        this.instance = axios.create({
-            baseURL: options.URL,
-            timeout: options.timeout,
-            headers: {
-                'x-api-key': options.ApiKey
-            }
-        });
-    }
+    constructor(engine?: constructor<unknown>) {
+        const childContainer = container.createChildContainer();
 
+        if(engine) {
+            childContainer.register('Engine', { useClass: engine });
+        } else {
+            childContainer.register('Engine', { useClass:  IPFSEngine });
+        }
 
-    /**
-     * Save file in a specific storage engine
-     * @param {object} options - File to save and additional parameters
-     * @returns {string} returns the file identifier or location
-     * @throws {BadRequestError} File was not save on IPFS
-     * @throws {InternalServerError} Internal server error
-     */
-    async saveFile(options: object): Promise<string> {
-        const body = options as IIpfsSave;
-        
-        const formData = new FormData();
-        
-        formData.append('file', body.file);
-        formData.append('address', body.address);
-        formData.append('fileName', body.fileName);
-        formData.append('keepOriginalName', String(body.keepOriginalName));
-
-        const res = await this.instance.post('/file', formData);
-        
-        return res.data.CID;
+        this.storageEngine = childContainer.resolve(Storage);
     }
 
     /**
-     * Get file from storage engine
-     * @param {object} options - File identifier or location and additional parameters
-     * @returns {string} returns the file
-     * @throws {NotFoundError} File not found in IPFS
-     * @throws {InternalServerError} Internal server error
+     * Get storage engine instance
+     * @returns {IStorageEngine} return storage engine instance
      */
-    async getFile(options: object): Promise<string> {
-        const params = options as IIpfsDocument;
-        
-        const file = await this.instance.get(`/file?address=${params.address}&cid=${params.cid}`);
-        
-        return file.data.file;
-    }
-    
-    /**
-     * Update file stored
-     * @param {object} options - File identifier or location and additional parameters
-     * @throws {NotFoundError} File not found in IPFS or not registered in smart contract
-     * @throws {ForbiddenError} User is not the owner of the file
-     * @throws {InternalServerError} Internal server error
-     */
-    async updateFile(options: object): Promise<void> {
-        const body = options as IIpfsUpdate;
-
-        const formData = new FormData();
-
-        const serverHash = await this.getChallenge(body.address);
-
-        const signature = this.generateSignature(serverHash, body.privateKey);
-
-        formData.append('file', body.file);
-        formData.append('address', body.address);
-        formData.append('cid', body.cid || '');
-        formData.append('sigV', signature.sigV.toString());
-        formData.append('sigR', signature.sigR);
-        formData.append('sigS', signature.sigS);
-        formData.append('hash', serverHash);
-
-        await this.instance.put('/file', formData);
-    }
-
-    /**
-     * Delete file from storage engine
-     * @param {object} options - Identifier of the file to delete and additional parameters
-     * @throws {NotFoundError} File not found in IPFS or not registered in smart contract
-     * @throws {ForbiddenError} User is not the owner of the file
-     * @throws {InternalServerError} Internal server error
-     */
-    async deleteFile(options: object): Promise<void> {
-        const params = options as IIpfsDelete;
-
-        const serverHash = await this.getChallenge(params.address);
-
-        const signature = this.generateSignature(serverHash, params.privateKey);
-
-        const body = {
-            address: params.address,
-            cid: params.cid,
-            hash: serverHash,
-            sigV: signature.sigV,
-            sigR: signature.sigR,
-            sigS: signature.sigS
-        };
-
-        await this.instance.delete('/file', {data : body});
-    }
-
-    /**
-     * Get hash by user address
-     * @param {string} address - User address
-     * @returns {Promise<string>} returns hash made by user address
-     */
-    private async getChallenge (address: string): Promise<string> {
-        const res = await this.instance.get(`/challenge?address=${address}`);
-
-        return res.data.hash;
-    }
-
-    /**
-     * Generate signature from hash
-     * @param {string} serverHash - Hash
-     * @param {string} privateKey - Private key
-     * @returns {IIpfsSignature} Signature
-     */
-    private generateSignature(serverHash: string, privateKey: string): IIpfsSignature {
-        const hash = Buffer.from(serverHash, 'hex');
-        const privateKeyBuffer = Buffer.from(privateKey, 'hex');
-        const { v, r, s } = ecsign(hash, privateKeyBuffer);
-
-        const signData: IIpfsSignature = {
-            hash: serverHash,
-            sigR: r.toString('hex'),
-            sigS: s.toString('hex'),
-            sigV: v
-        };
-
-        return signData;
+    getStorageEngine (): IStorageEngine {
+        return this.storageEngine;
     }
 }
