+++ src/storageEngine/index.ts
@@ -1,7 +1,9 @@
-import StorageEngine from './StorageEngine';
+import Storage from './Storage';
 import IStorageEngine from './IStorageEngine';
+import StorageEngine from './StorageEngine';
 
 export {
     StorageEngine,
+    Storage,
     IStorageEngine
 };
